+++ v2/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 LIFO documentation style
 
+## V2.2
+
+- ExecCmd is DEPRECATED, to be replaced by ExecNoOutputCmd
+- New functions
+    - ExecNoOutputCmd(cmdAndArgs string) error
+    - ExecOutputCmd(cmdAndArgs string) ([]byte, error)
+    - IsExistingFile(filename string) bool
+    - IsExistingPlainFile(filename string) bool
 ## V2.1
 
 - ExecutableReachableByPath(cmd...) error